+++ src/main/java/cz/cvut/fel/aic/agentpolis/config/AgentpolisConfig.java
@@ -19,6 +19,8 @@ public class AgentpolisConfig implements GeneratedConfig {
 
   public Integer maxVehicleSpeedInMeters;
 
+  public Boolean useMeasuredSpeed;
+
   public Visio visio;
 
   public String mapNodesFilepath;
@@ -45,6 +47,7 @@ public class AgentpolisConfig implements GeneratedConfig {
     this.srid = (Integer) agentpolisConfig.get("srid");
     this.pathToSaveRecordings = (String) agentpolisConfig.get("path_to_save_recordings");
     this.maxVehicleSpeedInMeters = (Integer) agentpolisConfig.get("max_vehicle_speed_in_meters");
+    this.useMeasuredSpeed = (Boolean) agentpolisConfig.get("use_measured_speed");
     this.visio = new Visio((Map) agentpolisConfig.get("visio"));
     this.mapNodesFilepath = (String) agentpolisConfig.get("map_nodes_filepath");
     this.simulationDuration = new SimulationDuration((Map) agentpolisConfig.get("simulation_duration"));