+++ src/test/java/cz/cvut/fel/aic/agentpolis/simmodel/environment/model/congestion/support/mock/TestCongestionModel.java
@@ -23,6 +23,7 @@ import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import cz.cvut.fel.aic.agentpolis.config.AgentpolisConfig;
 import cz.cvut.fel.aic.agentpolis.siminfrastructure.time.TimeProvider;
+import cz.cvut.fel.aic.agentpolis.simmodel.MoveUtil;
 import cz.cvut.fel.aic.agentpolis.simmodel.environment.congestion.CongestionModel;
 import cz.cvut.fel.aic.agentpolis.simmodel.environment.congestion.Connection;
 import cz.cvut.fel.aic.agentpolis.simmodel.environment.congestion.Link;
@@ -41,10 +42,16 @@ import java.security.ProviderException;
 public class TestCongestionModel extends CongestionModel {
 
 	@Inject
-	public TestCongestionModel(TransportNetworks transportNetworks, AgentpolisConfig config,
-							   SimulationProvider simulationProvider, TimeProvider timeProvider, ShapeUtils shapeUtils) throws ModelConstructionFailedException,
+	public TestCongestionModel(
+			TransportNetworks transportNetworks, 
+			AgentpolisConfig config,
+			SimulationProvider simulationProvider, 
+			TimeProvider timeProvider, 
+			ShapeUtils shapeUtils,
+			MoveUtil moveUtil
+	) throws ModelConstructionFailedException,
 			ProviderException {
-		super(transportNetworks, config, simulationProvider, timeProvider, shapeUtils);
+		super(transportNetworks, config, simulationProvider, timeProvider, shapeUtils, moveUtil);
 	}
 
 	public Connection getConnectionByNode(SimulationNode node) {