+++ CHANGELOG.md
@@ -27,6 +27,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 - `Trip` coordinated now have to be WKT printable
 - Speed is now stored internally in the original units (km, miles), another property `maxSpeedUnit` was added to `Edge` to store the unit type.
 - `EventProcessor` in `provideTimeProvider` method mad nullable to support easier injection for scripts that does not need the simulation, but need the travel time providers. It should be reverted as soon as Issue #54 will be resolved.
+- some `MoveUtil` static methods changed to instance methods (they need to be configurable)
 
 ## Added
 - `Trip` WKT export